move the servers into a dedicated folder
