+++ front/lib/actions/mcp_internal_actions/index.ts
@@ -7,9 +7,9 @@ import {
   getInternalMCPServerNameAndWorkspaceId,
   INTERNAL_MCP_SERVERS,
 } from "@app/lib/actions/mcp_internal_actions/constants";
-import { default as dataSourceUtilsServer } from "@app/lib/actions/mcp_internal_actions/data_source_utils";
-import { default as helloWorldServer } from "@app/lib/actions/mcp_internal_actions/helloworld";
-import { default as tableUtilsServer } from "@app/lib/actions/mcp_internal_actions/table_utils";
+import { default as dataSourceUtilsServer } from "@app/lib/actions/mcp_internal_actions/servers/data_source_utils";
+import { default as helloWorldServer } from "@app/lib/actions/mcp_internal_actions/servers/helloworld";
+import { default as tableUtilsServer } from "@app/lib/actions/mcp_internal_actions/servers/table_utils";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
 import { assertNever } from "@app/types";
