+++ front/lib/actions/mcp_internal_actions/servers/table_utils.ts
